fix fino al parsing delle capacità irrazionali, manca gestire il flusso e la correttezza dell'algoritmo

+++ src/util/scale_rationals.hpp
@@ -4,5 +4,5 @@
 #include "fraction.hpp"
 
 // Restituisce: (vector<int> scaled, int scaling_factor)
-std::pair<std::vector<int>, int>
+std::pair<std::vector<unsigned long long>, unsigned long long>
 scale_rationals(const std::vector<Fraction>& values);
